requirements.txt
Flask==3.1.0
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.165.0
pandas==2.2.3
requests==2.32.3
numpy==1.23.5